Separated FromType<> into own partial class file
