+++ DbMocker/MockTableImport.cs
@@ -10,8 +10,6 @@ namespace Apps72.Dev.Data.DbMocker
     {
         private static readonly string[] SPLIT_NEWLINE = new string[] { Environment.NewLine };
 
-        public const BindingFlags DefaultFromTypeBindingFlags = BindingFlags.Instance | BindingFlags.GetProperty | BindingFlags.Public;
-
         /// <summary />
         public static MockTable FromCsv(string content, string delimiter, CsvImportOptions options)
         {
@@ -78,38 +76,6 @@ namespace Apps72.Dev.Data.DbMocker
         public static MockTable FromFixed(Assembly assembly, string resourceName)
         {
             return new MockTableImportFixed(assembly, resourceName).GetMockTable();
-        }
-
-        /// <summary />
-        public static MockTable FromType<T>(
-            IEnumerable<T> rows = null,
-            BindingFlags propertyBindingFlags = DefaultFromTypeBindingFlags
-            )
-        {
-            var columns = new List<(string, Type)>();
-
-            var propertyInfos = typeof(T).GetProperties(propertyBindingFlags);
-
-            foreach (var propertyInfo in propertyInfos)
-            {
-                columns.Add((propertyInfo.Name, propertyInfo.PropertyType));
-            }
-
-            var table = WithColumns(columns.ToArray());
-
-            if (rows != null)
-            {
-                foreach (var row in rows)
-                {
-                    var values = propertyInfos
-                        .Select(pi => pi.GetValue(row))
-                        .ToArray();
-
-                    table.AddRow(values);
-                }
-            }
-
-            return table;
         }
 
         /// <summary />
